Add Java solution for Distinct_Numbers

+++ Sorting_and_Searching/Distinct_Numbers/Distinct_Numbers.java
@@ -16,9 +16,9 @@ import java.util.*;
 public class Main
 {
     public static int func(int n,int[] a){
-        Set<Integer> b=new HashSet<>();
+        Set<Long> b=new HashSet<>();
         for(int i:a){
-            b.add(i);
+            b.add((long) i);
         }
         return b.size();
     }
